+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 * [UICIRCLOG-115](https://issues.folio.org/browse/UICIRCLOG-115) Implement EXACT search for item barcode in circulation log
+* [UICIRCLOG-122](https://issues.folio.org/browse/UICIRCLOG-122) Make patron and staff information visible and searchable in Circulation log
 
 ## [3.0.0](https://github.com/folio-org/ui-circulation-log/tree/v3.0.0) (2022-02-24)
 [Full Changelog](https://github.com/folio-org/ui-circulation-log/compare/v2.3.0...v3.0.0)